+++ server/src/main/java/org/realityforge/gwt/appcache/server/BindingProperty.java
@@ -46,6 +46,32 @@ public final class BindingProperty
   }
 
   @Override
+  public boolean equals( final Object o )
+  {
+    if ( this == o )
+    {
+      return true;
+    }
+    else if ( o == null || getClass() != o.getClass() )
+    {
+      return false;
+    }
+    else
+    {
+      final BindingProperty that = (BindingProperty) o;
+      return _name.equals( that._name ) && _value.equals( that._value );
+    }
+  }
+
+  @Override
+  public int hashCode()
+  {
+    int result = _name.hashCode();
+    result = 31 * result + _value.hashCode();
+    return result;
+  }
+
+  @Override
   public String toString()
   {
     return "BindingProperty[name=" + _name + ", value=" + _value + "]";